fix(importer): read byte fields as signed
A byte field came back through the unsigned read used for raw stream bytes, so
a value of -1 arrived as 255. Short already sign extended, which is what makes
the asymmetry a slip rather than a decision.

No field in any distributed file is a byte, so no import changes: the whole
installation reimports byte for byte identical. The reader runs on third party
files as well, and a silently wrong number there is worse than a loud failure.

The existing width test passed a positive byte and could not catch it.

+++ importer/src/main/kotlin/org/openfoot/importer/SerializedStream.kt
@@ -132,7 +132,7 @@ class SerializedStreamReader(private val bytes: ByteArray) {
     private fun readFieldValue(typeCode: Char): Any? = when (typeCode) {
         'I' -> readInt()
         'Z' -> readByte() != 0
-        'B' -> readByte()
+        'B' -> readByte().toByte().toInt()
         'S' -> readSignedShort()
         'J' -> readLong()
         'D' -> Double.fromBits(readLong())
